requirements.txt
fastmcp>=2.12.0
uvicorn>=0.35.0
garminconnect>=0.2.38
garth>=0.5.17,<0.6.0
python-dotenv>=1.0.0
requests>=2.32.0
